// property-hub-backend/src/modules/commission-managers/commission-managers.dto.ts
import { IsEmail, IsString, IsNotEmpty, IsOptional } from 'class-validator';

export class CreateCommissionManagerDto {
    @IsEmail()
    email: string;

    @IsString()
    @IsNotEmpty()
    firstName: string;

    @IsString()
    @IsNotEmpty()
    lastName: string;

    @IsString()
    @IsOptional()
    phone?: string;
}

export class CommissionManagerDto {
    id: string;
    keycloakId: string | null;
    name: string;
    email: string;
    phone: string | null;
    role: string;
    status: string;
    createdAt: Date;
    updatedAt: Date;
}